backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
boto3==1.34.84
python-jose>=3.4.0
redis==5.0.3
python-multipart==0.0.9
pydantic[email]==1.10.7
azure-storage-blob==12.10.0
pytest==7.4.4
httpx==0.27.0
passlib[bcrypt]
